+++ tests/Ambev.DeveloperEvaluation.Unit/Domain/Entities/UserEntity/TestData/UserEntityTestData.cs
@@ -2,14 +2,14 @@ using Ambev.DeveloperEvaluation.Domain.Entities;
 using Ambev.DeveloperEvaluation.Domain.Enums;
 using Bogus;
 
-namespace Ambev.DeveloperEvaluation.Unit.Domain.Entities.TestData;
+namespace Ambev.DeveloperEvaluation.Unit.Domain.Entities.UserEntity.TestData;
 
 /// <summary>
 /// Provides methods for generating test data using the Bogus library.
 /// This class centralizes all test data generation to ensure consistency
 /// across test cases and provide both valid and invalid data scenarios.
 /// </summary>
-public static class UserTestData
+public static class UserEntityTestData
 {
     /// <summary>
     /// Configures the Faker to generate valid User entities.
